+++ CHANGELOG.md
@@ -26,7 +26,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - feature/1200030326491309 - Add ToArray and ArrayFactoryExceptions
 - feature/1200005256502350 - Add ToArrayInterface and ArrayFactoryInterface
 - feature/1199522825882943 - Add Symfony Var Dumper (https://github.com/symfony/var-dumper)
-
+- feature/1200636482135423 - Enhance ComparisonOperatorHelper
 
 ### Changed